Move user, group, role and permission related components to src/components/rbac/ (#2247)
